backend/requirements.txt
flask
flask-cors
transformers
torch
sentencepiece
PyPDF2
python-docx
reportlab
youtube-transcript-api
gunicorn